Enhance model management with version tracking and history; update related components for improved data handling
